[Fuzzy] Add API to create fuzzy terms automatically
If process is called and a variable has no terms, we fuzzy engine is
now creating default terms for it, covering all area between min and
max values. Width of each term is user defined.
If no width is defined, it is calculated by dividing the range
(max - min) by a default constant
If Id property is set, we expect that the variable is used as an id. So
the first and last item will be half the regular size used. This is
necessary to keep the distances for each term peak constant.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>