Change name getters API for variables and terms
Set a name size limit to setters, and change getters to write
to a buffer passed as arg instead of returning char*

Since fuzzy saves names only in fuzzylite vars, it must to
convert names to .c_str(), so it would need to allocate it
on memory to return to callers.

Before, it was just returning a pointer to this converted string,
making callers access freed memory.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>