ann+fuzzy: Create infra to save and load debug data
Add a method to save to a file all calls to functions that change
the state of the current engine and a method to load this file
and reproduce the saved changes.

This is used to reproduce a sml execution for debug purposes. All data
saved to inputs and outputs, predicted values and variables
adds/removals/disabled are saved.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>