Change project license from BSD-3 clause to Apache 2.0
The Soletta project chose to switch to Apache 2.0 to ensure that all
contributors explicitly grant the right to use any patents in their
contributions.  This avoids the possibility of a contributor
subsequently using patents to threaten the Soletta community or the
users and distributors of Soletta-based devices.  IoT is a rapidly
developing area, with many people filing patents and attempting to claim
territory; this change helps us maintain and protect our Open Source
project and community.

Other IoT projects have also chosen the Apache 2.0 license for similar
reasons, including IoTivity and the recently released Zephyr OS.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>